## Why is this change being made?
Adding SSI_QUIESCE_ASYNC case for DDR to handle shutdown notification and print "Goodbye world, DDR Manager ..."
## What is changed?

## How is the change being tested?

- [ ] No test, document update only
- [ X] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [X ] SVP
- [ ] FPGA (Add FPGA Config)
- [ X] Silicon System 19, node 2 - executed  'warm_start wsreset quiesce' command for testing/verification.
- [ ] Others, explain:

## Notes
![image.png](https://azurecsi.visualstudio.com/2610da6a-0168-48f8-b34a-f1c882cb6660/_apis/git/repositories/8fc5d0fc-7430-45b1-8963-5d075fba03da/pullRequests/156098/attachments/image.png)
## Risks

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature by integrating DDR Manager with SOS to register a shutdown callback.

#### PR Summary
The pull request adds a DDR driver framework interface that registers with the SOS startup/shutdown service to handle shutdown notifications, fulfilling the associated work item requirements.
- `src/libs/ddrss/ddr_manager/inc/ddr_manager_dfwk.h`: Adds new DDR interface structures and function prototypes.
- `src/libs/ddrss/ddr_manager/ddr_manager_dfwk.c`: Implements DDR Manager initialization, dispatch logic, and registers the shutdown callback with SOS.
- `src/libs/runtime_init/ddrss_init.c`: Introduces a new initialization component to invoke the DDR DFwk integration.
- `src/libs/ddrss/ddr_manager/CMakeLists.txt`: Updates the build configuration to include the new DDR DFwk module.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #3154263, #3160323

+++ src/services/startup_shutdown/startup_shutdown_thread.c
@@ -81,7 +81,7 @@ void sos_completion(PDFWK_ASYNC_REQUEST_HEADER request, void* p_completion_conte
 {
     FPFW_UNUSED(request);
     uint32_t interface_unique_flag = (uint32_t)p_completion_context;
-    SOS_LOG_TRACE("Request to an SSI %08lx completed (%x)", interface_unique_flag, (uintptr_t)request);
+    SOS_LOG_TRACE("Request to an SSI %08lx completed (%x)", (unsigned long)interface_unique_flag, (uintptr_t)request);
     int status = tx_event_flags_set(&s_sos_thread_ctx.ssi_flags, interface_unique_flag, TX_OR);
     FPFW_RUNTIME_ASSERT(status == TX_SUCCESS);
 }
@@ -180,7 +180,7 @@ void wait_ssi_complete(sos_stage_timeout_t current_stage)
         current_stage.timeout_ms = DEFAULT_SOS_TIMEOUT_MS;
     }
 
-    SOS_LOG_TRACE("SSI completion - Expected flags: %lx", s_sos_thread_ctx.expected_complete_flags);
+    SOS_LOG_TRACE("SSI completion - Expected flags: %lx", (unsigned long)s_sos_thread_ctx.expected_complete_flags);
     int status = tx_event_flags_get(&s_sos_thread_ctx.ssi_flags,
                                     s_sos_thread_ctx.expected_complete_flags,
                                     TX_AND_CLEAR,
